requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
werkzeug==2.0.1
pn532pi==1.0.4
smbus2==0.4.2